WIP - split off vector builtin functions
