+++ mathics/builtin/vectors/__init__.py
@@ -1,0 +1,13 @@
+"""
+Operations on Vectors
+
+In mathematics and physics, a vector is a term that refers colloquially to some quantities that cannot be expressed by a single number. It is also a row or column of a matrix.
+
+In computer science, it is an array datas structure consiting of collection of elements identified by at least on array index or key.
+
+In Mathics vectors as are Lists.  one never needs to distinguish between row and column vectors. As with other objects vectors can mix number and symbolic elements.
+
+Vectors can be long, dense, or sparse.
+
+Here you find functions on Vectors.
+"""
