+++ mathics/core/convert/mpmath.py
@@ -8,6 +8,7 @@ from mathics.core.atoms import (
     MachineReal,
     PrecisionReal,
 )
+from mathics.core.symbols import SymbolList
 
 
 @lru_cache(maxsize=1024)
@@ -27,3 +28,24 @@ def from_mpmath(value, prec=None):
         return Complex(real, imag)
     else:
         raise TypeError(type(value))
+
+
+def to_mpmath_matrix(data, **kwargs):
+    """
+    Convert a Mathics matrix to one that can be used by mpmath.
+    None is returned if we can't convert to a mpmath matrix.
+    """
+
+    def mpmath_matrix_data(m):
+        if not m.has_form("List", None):
+            return None
+        if not all(element.has_form("List", None) for element in m.elements):
+            return None
+        return [[str(item) for item in row.elements] for row in m.elements]
+
+    if not isinstance(data, list):
+        data = mpmath_matrix_data(data)
+    try:
+        return mpmath.matrix(data)
+    except (TypeError, AssertionError, ValueError):
+        return None
